simplify subs page and remove loaders
