+++ app/(app)/(home-pages)/p/[didOrHandle]/comments/page.tsx
@@ -36,15 +36,5 @@ async function ProfileCommentsLoader({ did }: { did: string }) {
 }
 
 function ProfileCommentsSkeleton() {
-  return (
-    <div className="flex flex-col gap-6 py-4">
-      {[0, 1, 2].map((i) => (
-        <div key={i} className="flex flex-col gap-2 animate-pulse">
-          <div className="h-4 w-48 bg-border-light rounded" />
-          <div className="h-3 w-full bg-border-light rounded" />
-          <div className="h-3 w-2/3 bg-border-light rounded" />
-        </div>
-      ))}
-    </div>
-  );
+  return null;
 }
