+++ app/(app)/lish/[did]/[publication]/dashboard/PublicationSubscribers.tsx
@@ -1,12 +1,17 @@
 "use client";
-import { usePublicationData } from "./PublicationSWRProvider";
 import { ButtonPrimary } from "components/Buttons";
 import { useSmoker } from "components/Toast";
 import { Separator } from "components/Layout";
 import { useLocalizedDate } from "src/hooks/useLocalizedDate";
-import { useDashboardState } from "components/PageLayouts/dashboardState";
+import {
+  useDashboardState,
+  useSetDashboardState,
+} from "components/PageLayouts/dashboardState";
 import { AtmosphereAccount } from "components/Icons/AtmosphereAccount";
 import { EmailTiny } from "components/Icons/EmailTiny";
+import { DashboardPageLayout } from "components/PageLayouts/DashboardPageLayout";
+import { Popover } from "components/Popover";
+import { Checkbox } from "components/Checkbox";
 
 export type SubscriberStatus = "subscribed" | "unconfirmed" | "unsubscribed";
 
@@ -19,75 +24,15 @@ export type MergedSubscriber = {
   status: SubscriberStatus;
 };
 
-export function useMergedSubscribers(): MergedSubscriber[] | null {
-  let { data: publication } = usePublicationData();
-  let { subscriberStatus } = useDashboardState();
-
-  if (!publication) return null;
-  // ATProto subscribers have no email lifecycle state — they're just present
-  // or absent, so they only count under the "subscribed" status filter.
-  let atprotoSubs = subscriberStatus.subscribed
-    ? publication.publication?.publication_subscriptions || []
-    : [];
-  let newsletterEnabled =
-    !!publication.publication?.publication_newsletter_settings?.enabled;
-  let emailSubs = newsletterEnabled
-    ? (publication.publication?.publication_email_subscribers || []).filter(
-        (s) => {
-          if (s.state === "confirmed") return subscriberStatus.subscribed;
-          if (s.state === "pending") return subscriberStatus.unconfirmed;
-          if (s.state === "unsubscribed") return subscriberStatus.unsubscribed;
-          return false;
-        },
-      )
-    : [];
-
-  let byDid = new Map<string, MergedSubscriber>();
-  let emailOnly: MergedSubscriber[] = [];
-  for (let s of atprotoSubs) {
-    let did = s.identities?.atp_did ?? undefined;
-    if (!did) continue;
-    byDid.set(did, {
-      key: `did:${did}`,
-      did,
-      handle: undefined,
-      email: undefined,
-      created_at: s.created_at,
-      status: "subscribed",
-    });
-  }
-  for (let s of emailSubs) {
-    let status: SubscriberStatus =
-      s.state === "pending"
-        ? "unconfirmed"
-        : s.state === "unsubscribed"
-          ? "unsubscribed"
-          : "subscribed";
-    let linkedDid = s.identities?.atp_did ?? undefined;
-    let existing = linkedDid ? byDid.get(linkedDid) : undefined;
-    if (existing && status === "subscribed") {
-      existing.email = s.email;
-      continue;
-    }
-    emailOnly.push({
-      key: `email:${s.id}`,
-      did: linkedDid,
-      handle: undefined,
-      email: s.email,
-      created_at: s.created_at,
-      status,
-    });
-  }
-  return [...byDid.values(), ...emailOnly];
-}
-
-export function SubscribersListView(props: {
+export function PublicationSubscribers(props: {
   subscribers: MergedSubscriber[];
   publicationShareUrl: string;
-  showPageBackground?: boolean;
+  publicationUri: string;
+  showPageBackground: boolean;
 }) {
   let smoker = useSmoker();
   let { subscriberStatus } = useDashboardState();
+  let filtered = props.subscribers.filter((s) => subscriberStatus[s.status]);
 
   let activeStatuses = (
     Object.keys(subscriberStatus) as SubscriberStatus[]
@@ -95,89 +40,83 @@ export function SubscribersListView(props: {
   let isDefaultStatusFilter =
     activeStatuses.length === 1 && activeStatuses[0] === "subscribed";
 
-  let filtered = props.subscribers.filter((s) => subscriberStatus[s.status]);
-
-  if (filtered.length === 0) {
-    if (!isDefaultStatusFilter) {
-      return (
-        <div
-          className={`italic text-tertiary
-            flex flex-col gap-0 text-center justify-center py-4 border rounded-md ${props.showPageBackground ? "border-border-light p-2" : "border-transparent"}`}
-          style={
-            props.showPageBackground
-              ? {
-                  backgroundColor:
-                    "rgba(var(--bg-page), var(--bg-page-alpha)) ",
-                }
-              : { backgroundColor: "transparent" }
-          }
-        >
-          <p className="font-bold">No subscribers match your filters!</p>
-        </div>
-      );
-    }
-    return (
-      <div
-        className={`italic text-tertiary flex flex-col gap-0 text-center justify-center py-4 border rounded-md ${props.showPageBackground ? "border-border-light p-2" : "border-transparent"}`}
-        style={
-          props.showPageBackground
-            ? {
-                backgroundColor: "rgba(var(--bg-page), var(--bg-page-alpha)) ",
-              }
-            : { backgroundColor: "transparent" }
-        }
-      >
-        <p className="font-bold"> No subscribers yet </p>
-        <p>Start sharing your publication!</p>
-        <ButtonPrimary
-          className="mx-auto mt-2"
-          onClick={(e) => {
-            e.preventDefault();
-            let rect = (e.currentTarget as Element)?.getBoundingClientRect();
-            navigator.clipboard.writeText(props.publicationShareUrl);
-            smoker({
-              position: {
-                x: rect ? rect.left + (rect.right - rect.left) / 2 : 0,
-                y: rect ? rect.top + 26 : 0,
-              },
-              text: "Copied Publication URL!",
-            });
-          }}
-        >
-          Copy Share Link
-        </ButtonPrimary>
-      </div>
-    );
-  }
+  let bgStyle = props.showPageBackground
+    ? { backgroundColor: "rgba(var(--bg-page), var(--bg-page-alpha)) " }
+    : { backgroundColor: "transparent" };
+  let bgBorder = props.showPageBackground
+    ? "border-border-light p-2"
+    : "border-transparent";
 
   return (
-    <div
-      className={`rounded-md ${props.showPageBackground ? "border-border-light p-2" : "border-transparent"}`}
-      style={
-        props.showPageBackground
-          ? {
-              backgroundColor: "rgba(var(--bg-page), var(--bg-page-alpha)) ",
-            }
-          : { backgroundColor: "transparent" }
+    <DashboardPageLayout
+      scrollKey={`dashboard-${props.publicationUri}-Subs`}
+      pageTitle="Subscribers"
+      mobileActions={<SubscriberStatusFilter />}
+      publication={props.publicationUri}
+      showHeader={true}
+      controls={
+        <div className="flex items-center justify-between gap-3 text-sm">
+          <div className="font-bold text-secondary px-1">
+            {filtered.length} Subscriber{filtered.length !== 1 && "s"}
+          </div>
+          <SubscriberStatusFilter />
+        </div>
       }
     >
-      <div className="subscriberListContent flex gap-2 flex-col ">
-        {filtered
-          .sort((a, b) => b.created_at.localeCompare(a.created_at))
-          .map((subscriber) => (
-            <div key={subscriber.key}>
-              <SubscriberListItem
-                handle={subscriber.handle}
-                did={subscriber.did}
-                email={subscriber.email}
-                createdAt={subscriber.created_at}
-                status={subscriber.status}
-              />
-              <hr className="border-border-light mt-2 last:hidden" />
-            </div>
-          ))}
-      </div>
-    </div>
+      {filtered.length === 0 ? (
+        <div
+          className={`italic text-tertiary flex flex-col gap-0 text-center justify-center py-4 border rounded-md ${bgBorder}`}
+          style={bgStyle}
+        >
+          {isDefaultStatusFilter ? (
+            <>
+              <p className="font-bold"> No subscribers yet </p>
+              <p>Start sharing your publication!</p>
+              <ButtonPrimary
+                className="mx-auto mt-2"
+                onClick={(e) => {
+                  e.preventDefault();
+                  let rect = (
+                    e.currentTarget as Element
+                  )?.getBoundingClientRect();
+                  navigator.clipboard.writeText(props.publicationShareUrl);
+                  smoker({
+                    position: {
+                      x: rect ? rect.left + (rect.right - rect.left) / 2 : 0,
+                      y: rect ? rect.top + 26 : 0,
+                    },
+                    text: "Copied Publication URL!",
+                  });
+                }}
+              >
+                Copy Share Link
+              </ButtonPrimary>
+            </>
+          ) : (
+            <p className="font-bold">No subscribers match your filters!</p>
+          )}
+        </div>
+      ) : (
+        <div className={`rounded-md ${bgBorder}`} style={bgStyle}>
+          <div className="subscriberListContent flex gap-2 flex-col ">
+            {filtered
+              .sort((a, b) => b.created_at.localeCompare(a.created_at))
+              .map((subscriber) => (
+                <div key={subscriber.key}>
+                  <SubscriberListItem
+                    handle={subscriber.handle}
+                    did={subscriber.did}
+                    email={subscriber.email}
+                    createdAt={subscriber.created_at}
+                    status={subscriber.status}
+                  />
+                  <hr className="border-border-light mt-2 last:hidden" />
+                </div>
+              ))}
+          </div>
+        </div>
+      )}
+    </DashboardPageLayout>
   );
 }
 
@@ -245,3 +184,63 @@ function SubscriberDate(props: { createdAt: string }) {
     </div>
   );
 }
+
+const SubscriberStatusFilter = () => {
+  let { subscriberStatus } = useDashboardState();
+  let setState = useSetDashboardState();
+  let count = Object.values(subscriberStatus).filter(Boolean).length;
+
+  return (
+    <Popover
+      className="text-sm px-2! py-1!"
+      trigger={
+        <div className="text-sm text-tertiary">
+          Filters {count > 0 && `(${count})`}
+        </div>
+      }
+    >
+      <Checkbox
+        small
+        checked={subscriberStatus.subscribed}
+        onChange={(e) =>
+          setState({
+            subscriberStatus: {
+              ...subscriberStatus,
+              subscribed: !!e.target.checked,
+            },
+          })
+        }
+      >
+        Subscribed
+      </Checkbox>
+      <Checkbox
+        small
+        checked={subscriberStatus.unconfirmed}
+        onChange={(e) =>
+          setState({
+            subscriberStatus: {
+              ...subscriberStatus,
+              unconfirmed: !!e.target.checked,
+            },
+          })
+        }
+      >
+        Unconfirmed
+      </Checkbox>
+      <Checkbox
+        small
+        checked={subscriberStatus.unsubscribed}
+        onChange={(e) =>
+          setState({
+            subscriberStatus: {
+              ...subscriberStatus,
+              unsubscribed: !!e.target.checked,
+            },
+          })
+        }
+      >
+        Unsubscribed
+      </Checkbox>
+    </Popover>
+  );
+};
